requirements.txt
flask==3.0.0
ollama==0.1.4
colorama==0.4.6
python-dotenv==1.0.0
markdown==3.5.1
pygments==2.17.2